+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Validate version constraints before generating the add-on manifest (Issue 38).
+
 ## [0.13.1] - 2025-02-27
 ### Fixed
 - Address deprecation warnings related to project usage at runtime.
@@ -248,6 +252,7 @@ Provided by the plugin:
  with a `ZapVersions.xml` from an add-on.
 
 
+[Unreleased]: https://github.com/zaproxy/gradle-plugin-add-on/compare/v0.13.1...HEAD
 [0.13.1]: https://github.com/zaproxy/gradle-plugin-add-on/compare/v0.13.0...v0.13.1
 [0.13.0]: https://github.com/zaproxy/gradle-plugin-add-on/compare/v0.12.0...v0.13.0
 [0.12.0]: https://github.com/zaproxy/gradle-plugin-add-on/compare/v0.11.0...v0.12.0